fix(asc-perf): hard-disable Studio background at runtime bootstrap
